Implemented AbstractWorkermanWebSocket
