+++ cli/workerman/data.php
@@ -6,12 +6,10 @@ use ChessServer\Command\CommandParser;
 use ChessServer\Command\Data\CommandContainer;
 use ChessServer\Command\Data\Db;
 use ChessServer\Socket\WorkermanClientStorage;
-use ChessServer\Socket\WorkermanWebSocket;
+use ChessServer\Socket\WorkermanDataWebSocket;
 use Dotenv\Dotenv;
 use Monolog\Logger;
 use Monolog\Handler\StreamHandler;
-use Workerman\Timer;
-use Workerman\Worker;
 
 require __DIR__  . '/../../vendor/autoload.php';
 
@@ -43,30 +41,6 @@ $context = [
     ],
 ];
 
-$server = (new WorkermanWebSocket($socketName, $context, $parser))->init($clientStorage);
-
-$worker = $server->getWorker();
-
-$worker->onWorkerStart = function(Worker $worker) use ($logger, &$db, $server) {
-    Timer::add(5, function() use ($logger, &$db, $server) {
-        try {
-            $db->getPdo()->getAttribute(\PDO::ATTR_SERVER_INFO);
-        } catch(\PDOException $e) {
-            try {
-                $db = new Db([
-                   'driver' => $_ENV['DB_DRIVER'],
-                   'host' => $_ENV['DB_HOST'],
-                   'database' => $_ENV['DB_DATABASE'],
-                   'username' => $_ENV['DB_USERNAME'],
-                   'password' => $_ENV['DB_PASSWORD'],
-                ]);
-                $server->setParser(new CommandParser(new CommandContainer($db)));
-                $logger->info('Successfully reconnected to Chess Data');
-            } catch(\PDOException $e) {
-                // Trying to connect to Chess Data...
-            }
-        }
-    });
-};
+$server = (new WorkermanDataWebSocket($socketName, $context, $parser))->init($clientStorage);
 
 $server->run();
