+++ src/Socket/AbstractWorkermanWebSocket.php
@@ -3,13 +3,12 @@
 namespace ChessServer\Socket;
 
 use ChessServer\Command\CommandParser;
-use ChessServer\Command\Game\LeaveCommand;
 use ChessServer\Exception\ParserException;
 use Workerman\Worker;
 
-class WorkermanWebSocket extends ChesslaBlabSocket
+class AbstractWorkermanWebSocket extends ChesslaBlabSocket
 {
-    private Worker $worker;
+    protected Worker $worker;
 
     public function __construct(string $socketName, array $context, CommandParser $parser)
     {
@@ -17,13 +16,11 @@ class WorkermanWebSocket extends ChesslaBlabSocket
 
         $this->worker = new Worker($socketName, $context);
         $this->worker->transport = 'ssl';
-
-        $this->connect()->message()->error()->close();
     }
 
-    public function getWorker()
+    public function run(): void
     {
-        return $this->worker;
+        $this->worker->runAll();
     }
 
     protected function connect()
@@ -85,33 +82,5 @@ class WorkermanWebSocket extends ChesslaBlabSocket
         };
 
         return $this;
-    }
-
-    protected function close()
-    {
-        $this->worker->onClose = function ($conn) {
-            if ($gameMode = $this->gameModeStorage->getById($conn->id)) {
-                $this->gameModeStorage->delete($gameMode);
-                $this->clientStorage->sendToMany($gameMode->getResourceIds(), [
-                    '/leave' => [
-                        'action' => LeaveCommand::ACTION_ACCEPT,
-                    ],
-                ]);
-            }
-
-            $this->clientStorage->detachById($conn->id);
-
-            $this->clientStorage->getLogger()->info('Closed connection', [
-                'id' => $conn->id,
-                'n' => $this->clientStorage->count()
-            ]);
-        };
-
-        return $this;
-    }
-
-    public function run()
-    {
-        $this->worker->runAll();
     }
 }
